refactor: move a few board method implementations
